+++ src/components/LoaderPage/MiniLoader.js
@@ -1,0 +1,7 @@
+import React from "react";
+import "./loaderpage.css";
+function MiniLoader() {
+  return <div className="lds-dual-ring"></div>;
+}
+
+export default MiniLoader;
