+++ src/components/LoaderPage/loaderpage.css
@@ -10,3 +10,27 @@
   height: 100%;
   background-color: rgb(255, 255, 255);
 }
+
+.lds-dual-ring {
+  display: inline-block;
+  width: 15px;
+  height: 15px;
+}
+.lds-dual-ring:after {
+  content: " ";
+  display: block;
+  width: 15px;
+  height: 15px;
+  border-radius: 50%;
+  border: 3px solid rgb(255, 3, 3);
+  border-color: rgb(255, 3, 3) transparent rgb(255, 3, 3) transparent;
+  animation: lds-dual-ring 1.2s linear infinite;
+}
+@keyframes lds-dual-ring {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
