+++ src/components/Product/product.css
@@ -30,7 +30,7 @@
   line-height: 2rem;
 }
 .card-btn {
-  width: 200px;
+  width: 100%;
 }
 .wishlistBtn {
   width: 200px !important;
